(chore) add clippy to candy machine (#490)
* add clippy to candy machine

* resync IDL to make it update

* fix clippy

* refactor tests and clippy stuff

* refactor tests again

* fix stuff
