+++ candy-machine/program/tests/utils/configs.rs
@@ -1,15 +1,18 @@
-use crate::CandyManager;
+use solana_program::pubkey::Pubkey;
+use solana_sdk::signer::Signer;
+
 use mpl_candy_machine::{
     CandyMachineData, Creator, EndSettings, GatekeeperConfig, HiddenSettings, WhitelistMintSettings,
 };
-use solana_program::pubkey::Pubkey;
-use solana_sdk::signer::Signer;
+
+use crate::CandyManager;
 
 pub const DEFAULT_UUID: &str = "ABCDEF";
 pub const DEFAULT_PRICE: u64 = 1e9 as u64;
 pub const ITEMS_AVAILABLE: u64 = 11;
 pub const DEFAULT_SYMBOL: &str = "SYMBOL";
 
+#[allow(dead_code)]
 pub fn quick_config(creator: Pubkey) -> CandyMachineData {
     custom_config(creator, None, true, true, None, None, None, None)
 }
