+++ candy-machine/program/tests/utils/mod.rs
@@ -1,6 +1,8 @@
+pub use candy_manager::*;
+pub use configs::*;
 pub use helper_transactions::*;
-pub use helpers::*;
-pub mod candy_manager;
-pub mod configs;
-pub mod helper_transactions;
+
+mod candy_manager;
+mod configs;
+mod helper_transactions;
 pub mod helpers;
